Refactor JWT service and update security configuration
The JWT service was refactored for better token handling and the addition of helper methods for token validation. The security configuration was also updated to improve request handling and the implementation of authorization. An authentic user's details retrieval mechanism was also introduced in the user controller. As a minor update, the version of spring-boot-starter-parent in the pom.xml file was bumped up to 3.2.1 for better compatibility and enhanced features.

Signed-off-by: Alyriana <128604938+Alyriana@users.noreply.github.com>