Refactor RentalController and add rental update functionality
The getRentalById method in the RentalController has been refactored to use ResponseEntity and the service layer directly. A new method, updateRental, has been added to the RentalService and RentalController to allow updating rental entities. In addition, an exception class named UploadFailureException was added to StorageService for image uploading failures. The MessageController has been extended with new functionalities. The SpringSecurityConfig was also updated to permit access to image endpoint.

Signed-off-by: Alyriana <128604938+Alyriana@users.noreply.github.com>