Refactor error handling and update several services
The major updates include refactoring the global exception handler and updating the JwtService, StorageService, RentalService, and UserService. The global exception handler was refactored to create a more generalized handling of exceptions. For JwtService, token validation has been improved, and for other services, methods for creating and updating items were modified. The changes are expected to increase efficiency and code maintainability.

Signed-off-by: Alyriana <128604938+Alyriana@users.noreply.github.com>